get users api config
- added get users api
- replace access token to utilise user's jwt
