+++ platform/src/core/apis/Account.js
@@ -1,5 +1,12 @@
-import { AUTH_URL, GOOGLE_AUTH_URL, LOGIN_URL } from '../urls/authentication';
+import { data } from 'autoprefixer';
+import { AUTH_URL, GOOGLE_AUTH_URL, LOGIN_URL, USERS_URL } from '../urls/authentication';
 import axios from 'axios';
+
+let jwtToken;
+if (typeof window !== 'undefined') {
+  jwtToken = window.localStorage.getItem('token');
+}
+axios.defaults.headers.common.Authorization = jwtToken;
 
 export const postUserCreationDetails = async (data) =>
   await axios.post(AUTH_URL, data).then((response) => response.data);
@@ -16,4 +23,8 @@ export const postUserLoginDetails = async (data) => {
     console.error(`Error posting user login details: ${error}`);
     throw error;
   }
+};
+
+export const getUserDetails = async (userID) => {
+  return await axios.get(`${USERS_URL}/${userID}`).then((response) => response.data);
 };
