+++ platform/src/core/urls/authentication.js
@@ -3,10 +3,12 @@ import { stripTrailingSlash } from '../utils/strings';
 
 const BASE_AUTH_URL = stripTrailingSlash(NEXT_PUBLIC_API_BASE_URL);
 
-export const AUTH_URL = `${BASE_AUTH_URL}/users?token=${NEXT_PUBLIC_API_TOKEN}`;
+export const AUTH_URL = `${BASE_AUTH_URL}/users`;
 
 export const GOOGLE_AUTH_URL = `${BASE_AUTH_URL}/users/auth/google`;
 
-export const LOGIN_URL = `${BASE_AUTH_URL}/users/loginUser?token=${NEXT_PUBLIC_API_TOKEN}`;
+export const LOGIN_URL = `${BASE_AUTH_URL}/users/loginUser`;
 
-export const UPDATE_PWD_URL = `${BASE_AUTH_URL}/users/updatePassword?token=${NEXT_PUBLIC_API_TOKEN}`;
+export const UPDATE_PWD_URL = `${BASE_AUTH_URL}/users/updatePassword`;
+
+export const USERS_URL = `${BASE_AUTH_URL}/users`;
